edit layout

+++ app/src/main/java/com/ajgroup/priceacsimulation/home/HomeFragment.kt
@@ -1,6 +1,7 @@
 package com.ajgroup.priceacsimulation.home
 
 import android.app.AlertDialog
+import android.content.DialogInterface
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
@@ -86,9 +87,6 @@ class HomeFragment : Fragment() {
             }
             dialog.show()
         }
-        binding.backButton.setOnClickListener {
-
-        }
     }
     private fun initRecyclerView() {
         binding.usersRecyclerView.layoutManager = LinearLayoutManager(this.context)
@@ -119,8 +117,24 @@ class HomeFragment : Fragment() {
                 }
                 dialog.show()
             },{
-                homeViewModel.deleteData(it)
-                Toast.makeText(context, "Delete Data Berhasil", Toast.LENGTH_SHORT).show()
+                val dialogWithButton = AlertDialog.Builder(context)
+                dialogWithButton.setTitle("Dialog Dengan Button")
+                dialogWithButton.setMessage("Dialog Dengan Button untuk berbagai Aksi")
+                dialogWithButton.setIcon(R.mipmap.ic_launcher_round)
+
+                dialogWithButton.setCancelable(false)
+                dialogWithButton.setPositiveButton("Yakin",
+                    DialogInterface.OnClickListener {
+                            dialogInterface, i ->
+                        homeViewModel.deleteData(it)
+                        Toast.makeText(context, "Data Berhasil di Delete", Toast.LENGTH_SHORT).show()
+                    })
+                dialogWithButton.setNegativeButton("Batal",
+                    DialogInterface.OnClickListener {
+                            dialogInterface, i ->
+                        Toast.makeText(context, "Batal Menghapus Data", Toast.LENGTH_SHORT).show() })
+                 dialogWithButton.show()
+                //Toast.makeText(context, "Delete Data Berhasil", Toast.LENGTH_SHORT).show()
             })
         })
